+++ src/hosts/CommercialNews.Api/Api/Admin/Contracts/Content/Tags/Responses/UpdateTagResponse.cs
@@ -10,7 +10,7 @@ namespace CommercialNews.Api.Api.Admin.Contracts.Content.Tags.Responses
 
         public bool IsActive { get; init; }
 
-        public int Version { get; init; }
+        public long Version { get; init; }
         public DateTime UpdatedAt { get; init; }
     }
 }
